+++ client/package.json
@@ -22,7 +22,8 @@
     "remark-gfm": "^1.0.0",
     "swr": "^0.5.6",
     "use-local-storage-state": "^10.0.0",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^1.1.2",
+    "yup": "^0.32.9"
   },
   "scripts": {
     "start": "craco start",
